Initial stab at refactoring in a similar suite to dsv
