--- lib/server.rb
@@ -1,4 +1,0 @@
-class Server
-  require 'server/server'
-  require 'server/secret'
-end
